+++ src/cjs.js
@@ -1,3 +1,5 @@
 const loader = require("./index");
 
 module.exports = loader.default;
+
+module.exports.defaultGetLocalIdent = require("./utils").defaultGetLocalIdent;